panthersearch-server/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
httpx==0.28.1
python-dotenv==1.2.2
beautifulsoup4==4.14.3
